[changed] Remove classnames dependency as a result of refactoring.
Address eslint warnings.

DRY refactor the code. Move all general logic into mixin.

`LinkMixin` is not a public API in `src/index.js`
therefore `getClassName()` method is not public too
and because it is used nowhere it has been removed.

And as a consequence `classnames` dependency has been removed also.